fix: some bug in BE

+++ backend/module/customer/customermodel/customer.go
@@ -11,7 +11,7 @@ type Customer struct {
 	Name     string                 `json:"name" gorm:"column:name;"`
 	Email    string                 `json:"email" gorm:"column:email;"`
 	Phone    string                 `json:"phone" gorm:"column:phone;"`
-	Point    float32                `json:"point" gorm:"column:point;"`
+	Point    int                    `json:"point" gorm:"column:point;"`
 	Invoices []invoicemodel.Invoice `json:"invoices,omitempty"`
 }
